+++ README.md
@@ -49,12 +49,7 @@ pip install -r requirements.txt
 
 This library is still in development and is not yet ready for use.
 
-An example script is provided that prints the activations of a GPT2-style Transformer using pretrained weights.
-To run it, simply execute the following command.
-
-```bash
-python example.py
-```
+The notebook(s) in the `examples` directory follow the tutorials on mechanistic interpretability provided [here](https://arena-ch1-transformers.streamlit.app).
 
 ## Relation to `TransformerLens`
